# elecciones2023-api/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"None","email":"none@none.com"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8081","description":"'@project.description@'"}],"paths":{"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/elecciones/secciones":{"get":{"tags":["elecciones-controller"],"operationId":"getSecciones","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}}},"/elecciones/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultados","parameters":[{"name":"seccionId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoResponseDTO"}}}}}}},"/elecciones/distritos":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritos","parameters":[{"name":"distrito_nombre","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Distrito"}}}}}}}},"/elecciones/cargos":{"get":{"tags":["elecciones-controller"],"operationId":"getCargosPorDistrito","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CargosResponseDTO"}}}}}}}},"components":{"schemas":{"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"SeccionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"}}},"ResultadoDTO":{"type":"object","properties":{"orden":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"votos":{"type":"integer","format":"int32"},"porcentaje":{"type":"number","format":"double"}}},"ResultadoResponseDTO":{"type":"object","properties":{"distrito":{"type":"string"},"seccion":{"type":"string"},"resultados":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDTO"}}}},"Distrito":{"type":"object","properties":{"distritoId":{"type":"integer","format":"int32"},"distritoNombre":{"type":"string"}}},"Cargo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"}}},"CargosResponseDTO":{"type":"object","properties":{"distrito":{"$ref":"#/components/schemas/DistritoDTO"},"cargos":{"type":"array","items":{"$ref":"#/components/schemas/Cargo"}}}},"DistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nombre":{"type":"string"}}}}}}