# elecciones2023-api/src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/DistritoDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class DistritoDTO {
    private int id;
    private String nombre;
}