# elecciones2023-api/src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/ResultadoResponseDTO.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ResultadoResponseDTO {
    private String distrito;
    private String seccion;
    private List<ResultadoDTO> resultados;
}