# elecciones2023-api/src/main/java/ar/edu/utn/frc/tup/lc/iv/models/Cargo.java
package ar.edu.utn.frc.tup.lc.iv.models;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class Cargo {
    private int id;
    private String nombre;
}