# elecciones2023-api/src/main/java/ar/edu/utn/frc/tup/lc/iv/models/Distrito.java
package ar.edu.utn.frc.tup.lc.iv.models;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class Distrito {
    private int distritoId;
    private String distritoNombre;
}